moved reg utils
