+++ scripts/regression/efficientnet_extractor.py
@@ -20,7 +20,7 @@ from spacer import config
 from spacer.tasks import extract_features
 from spacer.messages import DataLocation, ExtractFeaturesMsg
 
-from scripts.scripts_utils import build_traindata, start_training
+from .utils import build_traindata, start_training
 
 
 class ExtractFeatures:
